i am splitting out the index picker so i can test it and use it separately
